checker: Fix deletion of unused global variables
